Extract network settings to it's own component
